fix: drop lint-churn edits, fix https-guard blocked-response headers
The previous "linter" commit removed unrelated droppable warnings across 7
files to offset the warning ratchet. That backfired: touching those files
pulled their pre-existing warnings into the PR's changed-files set (2 -> 19),
introduced an ArrangeMissingParentheses regression in HlsManifestBuilder, and
its Headers change broke an EditMode test — while the ratchet itself is now
dominated by dev drift (baseline 12091 vs 12263 after the dev merge), which
these removals cannot meaningfully offset.

Revert all 7 droppable-file edits back to their original state, keeping only
the real fix in SimpleFetchApiImplementation:

- ShouldBlockNonHttps: keep `!string.IsNullOrEmpty(response.URL)` (Response is
  a struct whose URL can be null at runtime), which also clears the
  ConditionIsAlwaysTrue warning the guard originally introduced.
- Blocked response: leave Headers unset instead of `Headers = null`. The struct
  field defaults to null, so SimpleFetchApiSchemeGuardShould.
  BlockFileFinalUrlOutsideLocalSceneDevelopment (asserts Headers Is.Null) passes
  again, and the CS8625 null-literal warning is gone.

The residual lint ratchet failure is stale-baseline / dev drift, to be handled
with the no-warning-ratchet label rather than hand-removing dev's warnings.

+++ Explorer/Assets/DCL/AvatarRendering/AvatarShape/Tests/EditMode/FinishAvatarMatricesCalculationSystemShould.cs
@@ -151,12 +151,12 @@ namespace DCL.AvatarRendering.AvatarShape.Tests
             float3x2 fromJob = jobWrapper.RemoteAvatarsWorldBounds[index];
             Bounds expected = AvatarCustomSkinningComponent.NewWithLocalBounds(localBounds).ToWorldBounds(avatarBase.transform);
 
-            Assert.That(fromJob.c0.x, Is.EqualTo(expected.center.x).Within(TOLERANCE));
-            Assert.That(fromJob.c0.y, Is.EqualTo(expected.center.y).Within(TOLERANCE));
-            Assert.That(fromJob.c0.z, Is.EqualTo(expected.center.z).Within(TOLERANCE));
-            Assert.That(fromJob.c1.x, Is.EqualTo(expected.extents.x).Within(TOLERANCE));
-            Assert.That(fromJob.c1.y, Is.EqualTo(expected.extents.y).Within(TOLERANCE));
-            Assert.That(fromJob.c1.z, Is.EqualTo(expected.extents.z).Within(TOLERANCE));
+            Assert.That((float)fromJob.c0.x, Is.EqualTo(expected.center.x).Within(TOLERANCE));
+            Assert.That((float)fromJob.c0.y, Is.EqualTo(expected.center.y).Within(TOLERANCE));
+            Assert.That((float)fromJob.c0.z, Is.EqualTo(expected.center.z).Within(TOLERANCE));
+            Assert.That((float)fromJob.c1.x, Is.EqualTo(expected.extents.x).Within(TOLERANCE));
+            Assert.That((float)fromJob.c1.y, Is.EqualTo(expected.extents.y).Within(TOLERANCE));
+            Assert.That((float)fromJob.c1.z, Is.EqualTo(expected.extents.z).Within(TOLERANCE));
         }
 
         /// <summary>
